# Optimize discarding in ScanForLinks