This gives a pretty significant speedup and drops us down to an allocation per
link.

    name        old time/op    new time/op    delta
    LinkScan-4    4.03µs ± 3%    1.26µs ±12%  -68.79%  (p=0.000 n=10+10)

    name        old alloc/op   new alloc/op   delta
    LinkScan-4      881B ± 0%      112B ± 0%  -87.29%  (p=0.002 n=8+10)

    name        old allocs/op  new allocs/op  delta
    LinkScan-4      25.0 ± 0%       1.0 ± 0%  -96.00%  (p=0.000 n=10+10)